fix: print 64-bit time_t with PRId64 on 32-bit targets
On armv7l/armhf/static builds long is 32-bit but time_t is 64-bit, so
passing a time_t to %ld desynchronises va_arg and a later %s can be
handed an integer as a pointer (the STREAM:18 crash class, #23397).
Crash-capable sites: the dbengine page-conflict warning, extent read
errors, invalid replay timestamps, and a journal mmap-failure log.

Use %" PRId64 " with an (int64_t) cast for every time_t, %zu for
size_t/Word_t. Format strings only; output on 64-bit is unchanged.
An i386 _TIME_BITS=64 build goes from 145 -Wformat mismatches to 0.

+++ src/database/engine/pdc.c
@@ -1066,12 +1066,12 @@ static void epdl_extent_loading_error_log(struct rrdengine_instance *ctx, EPDL *
     nd_log_limit_static_global_var(erl, 1, 0);
     nd_log_limit(&erl, NDLS_DAEMON, priority,
                 "DBENGINE: error while reading extent from datafile %u of tier %d, at offset %" PRIu64 " (%u bytes) "
-                "%s from %ld (%s) to %ld (%s) %s%s: "
+                "%s from %" PRId64 " (%s) to %" PRId64 " (%s) %s%s: "
                 "%s",
                 epdl->datafile->fileno, ctx->config.tier,
                 BLOCK_TO_OFFSET(epdl->extent_block), epdl->extent_size,
                 used_epdl ? "to extract page (PD)" : used_descr ? "expected page (DESCR)" : "part of a query (PDC)",
-                start_time_s, start_time_str, end_time_s, end_time_str,
+                (int64_t)start_time_s, start_time_str, (int64_t)end_time_s, end_time_str,
                 used_epdl || used_descr ? " of metric " : "",
                 used_epdl || used_descr ? uuid : "",
                 msg);
